+++ Kinect-2/utils.h
@@ -18,7 +18,7 @@
 #define __KINECT_TRACK_IR_UTILS__
 void handleError(HRESULT hr, int critical, char* msg, ...);
 
-void getFaceRotationInDegrees(const Vector4* pQuaternion, int* pYaw, int* pPitch, int* pRoll);
+void getFaceRotationInDegrees(const Vector4* pQuaternion, int64_t* pYaw, int64_t* pPitch, int64_t* pRoll);
 
 // Safe release for interfaces
 template<class Interface>
